+++ src/main/java/com/ketheroth/uncrafter/Uncrafter.java
@@ -7,13 +7,18 @@ import com.ketheroth.uncrafter.core.registry.UncrafterContainerTypes;
 import net.minecraft.client.gui.screens.MenuScreens;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.eventbus.api.IEventBus;
+import net.minecraftforge.fml.InterModComms;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.config.ModConfig;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
+import net.minecraftforge.fml.event.lifecycle.InterModEnqueueEvent;
+import net.minecraftforge.fml.event.lifecycle.InterModProcessEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+
+import java.util.List;
 
 @Mod(Uncrafter.MODID)
 public class Uncrafter {
@@ -24,10 +29,9 @@ public class Uncrafter {
 	public Uncrafter() {
 		final IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
 
-//		modEventBus.addListener(this::setup);
 		modEventBus.addListener(this::clientSetup);
-//		modEventBus.addListener(this::enqueueIMC);
-//		modEventBus.addListener(this::processIMC);
+		modEventBus.addListener(this::enqueueIMC);
+		modEventBus.addListener(this::processIMC);
 
 		ModLoadingContext.get().registerConfig(ModConfig.Type.SERVER, Configuration.CONFIG);
 
@@ -38,27 +42,24 @@ public class Uncrafter {
 		MinecraftForge.EVENT_BUS.register(this);
 	}
 
-//	private void setup(final FMLCommonSetupEvent event) {
-//
-//	}
-
 	private void clientSetup(final FMLClientSetupEvent event) {
 		MenuScreens.register(UncrafterContainerTypes.UNCRAFTER_CONTAINER.get(), UncrafterScreen::new);
 	}
 
-//	private void enqueueIMC(final InterModEnqueueEvent event) {
-//		// some example code to dispatch IMC to another mod
-//		InterModComms.sendTo("examplemod", "helloworld", () -> {
-//			LOGGER.info("Hello world from the MDK");
-//			return "Hello world";
-//		});
-//	}
-//
-//	private void processIMC(final InterModProcessEvent event) {
-//		// some example code to receive and process InterModComms from other mods
-//		LOGGER.info("Got IMC {}", event.getIMCStream().
-//				map(m -> m.messageSupplier().get()).
-//				collect(Collectors.toList()));
-//	}
+	private void enqueueIMC(final InterModEnqueueEvent event) {
+		InterModComms.sendTo("uncrafter", "blacklistedRecipes", () -> List.of("minecraft:iron_block"));
+	}
+
+	private void processIMC(final InterModProcessEvent event) {
+		event.getIMCStream().filter(message -> message.method().equals("blacklistedRecipes")).forEach(message -> {
+			try {
+				List<?> recipes = (List<?>) message.messageSupplier().get();
+				recipes.stream().map(String.class::cast).forEach(recipe -> Configuration.IMC_BLACKLIST.add(recipe));
+			} catch (ClassCastException e) {
+				LOGGER.error("Error receiving IMC from : " + message.modId());
+			}
+			Configuration.IMC_BLACKLIST.forEach(System.out::println);
+		});
+	}
 
 }