+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 The default blacklist has some recipes to avoid item duplication.
 If you find more recipe that lead to duplication, please tell me, so I can add them to the default blacklist.
 - A server configuration to change the max amount of item that can be retrieved while uncrafting an item.
+- Inter Mod Communication. A mod can add a list of blacklisted recipes to the uncrafter.
 
 ## Changed
 - The recipe is now easier